chore: Making it more managable to inspect blocks
